requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
joblib==1.4.2
lxml==5.3.0
numpy==1.26.4
ollama>=0.6.0
openai>=1.68.0
openpyxl==3.1.5
pydantic==2.10.4
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
typer==0.15.1
uvicorn==0.34.0
xlrd==2.0.1
